DREES-106: Adding documentation and tests
Adding and correcting tests.

Correcting test on DcirBiologyActs extractor

PractitionnerClaimSpecialitySuite Modification for codecov

Removing fromRow, changed NgapActSuite

An error was corrected in MedicalActSuite

Correction on the NgapActSuite test

Correction on the order of the imports. Removed blanked lines unneeded.

Taking into account reviews on the PR.

Add comments, refacto on camelCase class name syntax.

Taking into account reviews on the PR.

- Adding new Ngap events for Ssr for future utilisation
- Adding comments and documentation
- Refacto on variables and class for syntax issue

Adding Documentation and tests

Adding Documentation and tests

This is a combination of 3 commits.
This is the 1st commit message:

Taking into account reviews on the PR.

Add comments, refacto on camelCase class name syntax.

This is the commit message #2:

Taking into account reviews on the PR.

- Adding new Ngap events for Ssr for future utilisation
- Adding comments and documentation
- Refacto on variables and class for syntax issue

This is the commit message #3:

Adding documentation and tests

Adding the definition of an Ngap act.

Adding documentation and tests

Correcting McoNgapActExtractor test

Taking into account the review

Better syntax and update on the groupID extraction for the DcirNgapActExtractor.
Tests updated.

Deleting a file which is not supposed to be in study but in etl/extractors

Added coments on the Mco CE Ngap extractor.

Adding test for SsrCEMedicalActExtractor

Adding documentation and tests

Adding the definition of an Ngap act.

Adding documentation and tests

Correcting McoNgapActExtractor test

Taking into account the review

Better syntax and update on the groupID extraction for the DcirNgapActExtractor.
Tests updated.

Deleting a file which is not supposed to be in study but in etl/extractors

Added coments on the Mco CE Ngap extractor.

Adding test for SsrCEMedicalActExtractor

Refacto name of class NgapActConfig